pyproject.toml
[tool.poetry]
name = "dask-ms"
version = "0.2.18"
description = "xarray Dataset from CASA Tables"
authors = ["Simon Perkins <simon.perkins@gmail.com>"]
license = "BSD3"
readme = "README.rst"
packages = [{include = "daskms"}]

[tool.poetry.dependencies]
python = "^3.8"
arcae = {version = "^0.2.1", optional = true}
appdirs = "^1.4.4"
dask = {extras = ["array"], version = "^2023.1.0"}
donfig = "^0.7.0"
python-casacore = "^3.5.1"
pyarrow = {version = "^13.0.0", optional=true}
zarr = {version = "^2.12.0", optional=true}
xarray = {version = "^2023.01.0", optional=true}
s3fs = {version = "^2023.1.0", optional=true}
minio = {version = "^7.1.11", optional=true}
pytest = {version = "^7.1.3", optional=true}

[tool.poetry.scripts]
dask-ms = "daskms.apps.entrypoint:main"
fragments = "daskms.apps.fragments:main"

[tool.poetry.extras]
v030 = ["arcae", "xarray"]
arrow = ["pyarrow"]
xarray = ["xarray"]
zarr = ["zarr"]
s3 = ["s3fs"]
complete = ["s3fs", "pyarrow", "xarray", "zarr"]
testing = ["minio", "pytest"]

[tool.poetry.group.dev.dependencies]
tbump = "^6.9.0"
pre-commit = "^2.20.0"
black = "^22.8.0"
mypy = "^1.6.1"

[tool.poetry.group.docs.dependencies]
furo = "^2022.9.15"
Sphinx = "^5.1.1"
numpydoc = "^1.4.0"
Pygments = "^2.13.0"
sphinx-copybutton = "^0.5.0"

[build-system]
requires = ["setuptools", "poetry-core"]
build-backend = "poetry.core.masonry.api"
